Version 3.07   * Prefetcher improvements, add global history buffer-based prefetcher   * HOOK_PERIODIC_INS: Instruction-based periodic callback   * Implement CLONE_CHILD_CLEARTID syscall interface   * Add example scripts for periodic statistics, periodic McPAT, simulating limited iteration counts   * Support for Pin 2.12   * Fixes to Python environment   * Various bugfixes

+++ common/core/memory_subsystem/cache/cache_base.cc
@@ -3,7 +3,7 @@
 
 CacheBase::CacheBase(String name, UInt32 cache_size, UInt32 associativity, UInt32 cache_block_size):
    m_name(name),
-   m_cache_size(k_KILO * cache_size),
+   m_cache_size(k_KILO * UInt64(cache_size)),
    m_associativity(associativity),
    m_blocksize(cache_block_size)
 {
